requirements.txt
absl-py==0.2.2
asn1crypto==0.24.0
astor==0.6.2
attrs==18.1.0
Automat==0.7.0
bleach==1.5.0
boto==2.48.0
boto3==1.7.39
botocore==1.10.39
bz2file==0.98
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
docker-py==1.10.6
docker-pycreds==0.3.0
docutils==0.14
gast==0.2.0
gensim==3.4.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jmespath==0.9.3
joblib==0.12.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.11
matplotlib==2.2.2
mecab-python3==0.7
numpy==1.14.3
parsel==1.4.0
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
queuelib==1.5.0
requests==2.18.4
retrying==1.3.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
scrapinghub==2.0.3
Scrapy==1.5.0
scrapy-splash==0.7.2
service-identity==17.0.0
shub==2.8.0
six==1.11.0
sklearn==0.0
smart-open==1.5.7
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.4
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
websocket-client==0.48.0
Werkzeug==0.14.1
zope.interface==4.5.0
